refactor: split body_model.py into anthropometrics, segments, and orchestration (issue #112)
Extract anthropometric tables/constants into body_anthropometrics.py and
segment XML helpers into body_segments.py; body_model.py becomes a thin
orchestration layer that re-exports the full public API unchanged.
